move init of env vars to fixtures.js, so we can run `TEST_WITH_FAKE_MONGO=1 npm run test:integration` locally without having to setup anything
